+++ tango/stick_filters/filters.go
@@ -19,6 +19,11 @@ func CreateFilters() map[string]stick.Filter {
 		"json_decode":       jsonDecode,
 		"boolify":           boolifyFilter,
 		"bool_switch":       boolSwitchFilter,
+		"exists":            existsFilter,
+		"value":             valueFilter,
+		"apply_mapping":     applyMappingFilter,
+		"fatality":          fatalityFilter,
+		"extract_objects":   extractObjectsFilter,
 	}
 }
 
@@ -105,4 +110,32 @@ func boolSwitchFilter(ctx stick.Context, val stick.Value, args ...stick.Value) s
 		return stickify(args[0])
 	}
 	return stickify(args[1])
+}
+
+// existsFilter checks if a variable exists in the current context
+func existsFilter(ctx stick.Context, val stick.Value, args ...stick.Value) stick.Value {
+	if (ctx == nil) || (ctx.Scope() == nil) {
+		return stickify(false)
+	}
+	var name = stick.CoerceString(val)
+	_, exists := ctx.Scope().All()[name]
+	return stickify(exists)
+}
+
+// valueFilter returns the value of a variable in the current scope and context,
+// or a default value if it does not exist
+func valueFilter(ctx stick.Context, val stick.Value, args ...stick.Value) stick.Value {
+	if (ctx == nil) || (ctx.Scope() == nil) {
+		return stickify(nil)
+	}
+	var name = stick.CoerceString(val)
+	defaultVal := stickify(nil)
+	if len(args) > 0 {
+		defaultVal = stickify(args[0])
+	}
+	value, exists := ctx.Scope().All()[name]
+	if exists {
+		return stickify(value)
+	}
+	return defaultVal
 }