+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] 
+
+## [1.6.0] - 2023-12-26
+
+### Added
+- `exists` filter to check if a variable with some name exists in the current scope
+- `value` filter to get the value of a variable specified by a name in the current scope or a default
+- `apply_mapping` filter to map objects or variables in scope to a new object
+- `extract_objects` filter to extract an array of objects from flat variables in the current scope
+- `fatality` filter to exit template processing with an error message
 
 ### Changed
 - Updated informatory texts